Vivaldi support, refactored code, updated version number, updated readme
