+++ Makefile
@@ -41,7 +41,7 @@ git-add:
 
 
 BEAT_VERSION=0.0.3
-VERSION_QUALIFIER=alpha1
+VERSION_QUALIFIER=alpha2
 
 VERSION=${BEAT_VERSION}-${VERSION_QUALIFIER}
 VERSION_GOFILEPATH=${BEAT_PATH}/version/version.go
